+++ GotYourBAC/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:value="@integer/google_play_services_version" />
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="AIzaSyAUIMMNWHg3Scvn3t_ItkDQfzwk3mOVqeI" />
+            android:value="@string/geoAPIkey" />
 
         <activity
             android:name="com.undergrads.ryan.MainActivity"
@@ -38,10 +38,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name="com.undergrads.ryan.SkeletonActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="portrait" />
+
         <activity
             android:name="com.undergrads.ryan.MenuActivity"
             android:label="@string/title_activity_menu" />
