+++ GotYourBAC/src/main/res/layout/fragment_pick_game_mode.xml
@@ -32,6 +32,7 @@
         <Button
             android:id="@+id/sign_out_button"
             android:layout_width="200dp"
+            android:layout_height="wrap_content"
             android:text="@string/sign_out" />
 
         <Button
