+++ GotYourBAC/src/main/res/layout/fragment_play_game.xml
@@ -154,16 +154,16 @@
 
     <TextView
         android:id="@+id/comment"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="TextView"
+        android:textSize="20sp"
+        android:visibility="invisible"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.665"
-        android:layout_width="106dp"
-        android:layout_height="33dp"
-        android:text="TextView"
-        android:textSize="20sp"
-        android:visibility="invisible"
         tools:layout_editor_absoluteX="139dp"
         tools:layout_editor_absoluteY="448dp" />
